handle more exceptions
